# golang: ~15x speedup for decodeTensor()