Make decodeTensor() faster by running binary.Read() for the whole slice in last dimension.

Similar to https://github.com/tensorflow/tensorflow/pull/14427

before:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/tensorflow/tensorflow/tensorflow/go
BenchmarkNewTensor/[150528]-8                200           7459717 ns/op
BenchmarkDecodeTensor/[150528]-8             100          11205557 ns/op
PASS
ok      github.com/tensorflow/tensorflow/tensorflow/go  3.447s


after:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/tensorflow/tensorflow/tensorflow/go
BenchmarkNewTensor/[150528]-8                200           7009254 ns/op
BenchmarkDecodeTensor/[150528]-8            2000            747224 ns/op
PASS
ok      github.com/tensorflow/tensorflow/tensorflow/go  3.793s